Prior to refactoring, creates a folder for UserPage components and its childs.
