+++ src/App.js
@@ -15,7 +15,7 @@ import SignupSuccess from './Components/Authentication/SignupSuccess'
 import PasswordReset from './Components/Authentication/PasswordReset'
 import ChangePassword from './Components/Authentication/ChangePassword'
 import PasswordResetSuccess from './Components/Authentication/PasswordResetSuccess'
-import UserPage from './Components/UserPage'
+import UserPage from './Components/UserPage/UserPage'
 import ScrollToTop from './Modules/ScrollToTop'
 import { Container, Sticky, Sidebar } from 'semantic-ui-react'
 import { connect } from 'react-redux'
