+++ src/Components/UserPage/UserPage.jsx
@@ -1,9 +1,9 @@
 import React, { Component } from 'react'
-import HostProfileForm from './HostProfile/HostProfileForm'
-import HostProfile from './HostProfile/HostProfile'
+import HostProfileForm from '../HostProfile/HostProfileForm'
+import HostProfile from '../HostProfile/HostProfile'
 import { connect } from 'react-redux'
 import { Header, Segment, Form, Dropdown, Button, Message, Divider, Image, Icon } from 'semantic-ui-react'
-import { LOCATION_OPTIONS } from '../Modules/locationData'
+import { LOCATION_OPTIONS } from '../../Modules/locationData'
 import axios from 'axios'
 import ReactAvatarEditor from 'react-avatar-editor'
 import Popup from 'reactjs-popup'
